+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>3.3.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>shop.S5G</groupId>
+	<groupId>shop.s5g</groupId>
 	<artifactId>bookApi</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>bookApi</name>
