+++ src/main/java/shop/s5g/bookApi/entity/BookAuthor.java
@@ -1,4 +1,4 @@
-package shop.S5G.bookApi.entity;
+package shop.s5g.bookApi.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,7 +7,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
