+++ src/main/java/shop/s5g/bookApi/entity/Profile.java
@@ -1,10 +1,9 @@
-package shop.S5G.bookApi.entity;
+package shop.s5g.bookApi.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -12,17 +11,18 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@NoArgsConstructor
 @AllArgsConstructor
-public class BookStatus {
+@Builder
+public class Profile {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long bookStatusId;
+    private long profileId;
 
-    private String typeName;
-
-    public BookStatus(String typeName) {
-        this.typeName = typeName;
-    }
+    private String birth;
+    private int debutYear;
+    private String introduction;
+    private String imageName;
+    private boolean active;
 }
