+++ src/main/java/shop/s5g/bookApi/entity/Publisher.java
@@ -1,11 +1,10 @@
-package shop.S5G.bookApi.entity;
+package shop.s5g.bookApi.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -13,16 +12,17 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class Profile {
+public class Publisher {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long profileId;
+    private long publisherId;
 
-    private String birth;
-    private int debutYear;
-    private String introduction;
-    private String imageName;
+    private String publisherName;
     private boolean active;
+
+    public Publisher(String publisherName, boolean active) {
+        this.publisherName = publisherName;
+        this.active = active;
+    }
 }
