+++ src/main/java/shop/s5g/bookApi/repository/AuthorRepository.java
@@ -1,9 +1,9 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.Author;
+import shop.s5g.bookApi.entity.Author;
 
 public interface AuthorRepository extends JpaRepository<Author, Long> {
     Optional<Author> findByName(String name);
