+++ src/main/java/shop/s5g/bookApi/repository/AuthorTypeRepository.java
@@ -1,9 +1,9 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.AuthorType;
+import shop.s5g.bookApi.entity.AuthorType;
 
 public interface AuthorTypeRepository extends JpaRepository<AuthorType, Long> {
