+++ src/main/java/shop/s5g/bookApi/repository/BookAuthorRepository.java
@@ -1,7 +1,7 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.BookAuthor;
+import shop.s5g.bookApi.entity.BookAuthor;
 
 public interface BookAuthorRepository extends JpaRepository<BookAuthor, Long> {
