+++ src/main/java/shop/s5g/bookApi/repository/BookCategoryRepository.java
@@ -1,8 +1,8 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.BookCategory;
-import shop.S5G.bookApi.entity.BookCategoryPk;
+import shop.s5g.bookApi.entity.BookCategory;
+import shop.s5g.bookApi.entity.BookCategoryPk;
 
 public interface BookCategoryRepository extends JpaRepository<BookCategory, BookCategoryPk> {
