+++ src/main/java/shop/s5g/bookApi/repository/BookRepository.java
@@ -1,7 +1,7 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.Book;
+import shop.s5g.bookApi.entity.Book;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
