+++ src/main/java/shop/s5g/bookApi/repository/BookStatusRepository.java
@@ -1,9 +1,9 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.BookStatus;
+import shop.s5g.bookApi.entity.BookStatus;
 
 public interface BookStatusRepository extends JpaRepository<BookStatus, Long> {
     Optional<BookStatus> findByTypeName(String typeName);
