+++ src/main/java/shop/s5g/bookApi/repository/CategoryRepository.java
@@ -1,9 +1,9 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.Category;
+import shop.s5g.bookApi.entity.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> findByCategoryNameAndParentCategory(String categoryName, Category parentCategory);
