+++ src/main/java/shop/s5g/bookApi/repository/ProfileRepository.java
@@ -1,7 +1,7 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.Profile;
+import shop.s5g.bookApi.entity.Profile;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long> {
