+++ src/main/java/shop/s5g/bookApi/repository/PublisherRepository.java
@@ -1,9 +1,9 @@
-package shop.S5G.bookApi.repository;
+package shop.s5g.bookApi.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.S5G.bookApi.entity.Publisher;
+import shop.s5g.bookApi.entity.Publisher;
 
 public interface PublisherRepository extends JpaRepository<Publisher, Long> {
     Optional<Publisher> findByPublisherName(String publisherName);
