+++ src/test/java/shop/s5g/bookApi/BookApiApplicationTests.java
@@ -1,4 +1,4 @@
-package shop.S5G.bookApi;
+package shop.s5g.bookApi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
